Switch to 4bpp mode
Lotsa hacks here:
I'm still using 8bpp functions (graphx and fontlibc), but I've changed the palettes of the sprites so that the upper nibble of the sprites equals the lower nibble, allowing things to be rendered at half-res horizontally. Everything also gets drawn twice, with an X offset, to fill in the gaps between 4bpp lines.